+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneDeviceEnrollmentLimitRestriction.Tests.ps1
@@ -40,8 +40,13 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             Mock -CommandName Remove-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
             }
+
             Mock -CommandName Update-DeviceConfigurationPolicyAssignment -MockWith {
             }
+
+            Mock -CommandName Invoke-MgGraphRequest -MockWith {
+            }
+
             # Mock Write-M365DSCHost to hide output during the tests
             Mock -CommandName Write-M365DSCHost -MockWith {
             }
@@ -53,14 +58,28 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name "When the restriction doesn't already exist" -Fixture {
             BeforeAll {
                 $testParams = @{
+                    Id          = '12345-12345-12345-12345-12345_Limit'
                     DisplayName = 'My DSC Restriction'
                     Ensure      = 'Present'
                     Credential  = $Credential
+                    Priority    = 1
                     Limit       = 15
                 }
 
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
                     return $null
+                }
+
+                Mock -CommandName New-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
+                    return @{
+                        AdditionalProperties = @{
+                            '@odata.type' = '#microsoft.graph.deviceEnrollmentLimitConfiguration'
+                            Limit         = 15
+                        }
+                        Id                   = '12345-12345-12345-12345-12345_Limit'
+                        Priority             = 1
+                        DisplayName          = 'My DSC Restriction';
+                    }
                 }
             }
 
@@ -81,20 +100,23 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name 'When the restriction already exists and is NOT in the Desired State' -Fixture {
             BeforeAll {
                 $testParams = @{
+                    Id          = '12345-12345-12345-12345-12345_Limit'
                     DisplayName = 'My DSC Restriction'
                     Ensure      = 'Present'
                     Credential  = $Credential
                     Limit       = 15
+                    Priority    = 1
                 }
 
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
                     return @{
                         AdditionalProperties = @{
                             '@odata.type' = '#microsoft.graph.deviceEnrollmentLimitConfiguration'
-                            Limit         = 12
+                            Limit         = 12 # Drift
                         }
-                        id                   = '12345-12345-12345-12345-12345'
-                        DisplayName          = 'My DSC Restriction'; #Drift
+                        Priority             = 1
+                        Id                   = '12345-12345-12345-12345-12345_Limit'
+                        DisplayName          = 'My DSC Restriction';
                     }
                 }
             }
@@ -116,10 +138,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
         Context -Name 'When the restriction already exists and IS in the Desired State' -Fixture {
             BeforeAll {
                 $testParams = @{
+                    Id          = '12345-12345-12345-12345-12345_Limit'
                     DisplayName = 'My DSC Restriction'
                     Ensure      = 'Present'
                     Credential  = $Credential
                     Limit       = 15
+                    Priority    = 1
                 }
 
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
@@ -128,7 +152,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             '@odata.type' = '#microsoft.graph.deviceEnrollmentLimitConfiguration'
                             Limit         = 15
                         }
-                        id                   = '12345-12345-12345-12345-12345'
+                        Id                   = '12345-12345-12345-12345-12345_Limit'
+                        Priority             = 1
                         DisplayName          = 'My DSC Restriction'
                     }
                 }
@@ -145,6 +170,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName = 'My DSC Restriction'
                     Ensure      = 'Absent'
                     Credential  = $Credential
+                    Id          = '12345-12345-12345-12345-12345_Limit'
+                    Priority    = 1
                     Limit       = 15
                 }
 
@@ -154,7 +181,8 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                             '@odata.type' = '#microsoft.graph.deviceEnrollmentLimitConfiguration'
                             Limit         = 12
                         }
-                        id                   = '12345-12345-12345-12345-12345'
+                        Id                   = '12345-12345-12345-12345-12345_Limit'
+                        Priority             = 1
                         DisplayName          = 'My DSC Restriction'
                     }
                 }
@@ -183,11 +211,13 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 }
 
                 Mock -CommandName Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -MockWith {
-                    return @{AdditionalProperties = @{
+                    return @{
+                        AdditionalProperties = @{
                             '@odata.type' = '#microsoft.graph.deviceEnrollmentLimitConfiguration'
                             Limit         = 12
                         }
-                        id                        = '12345-12345-12345-12345-12345'
+                        Id                        = '12345-12345-12345-12345-12345_Limit'
+                        Priority                  = 1
                         DisplayName               = 'My DSC Restriction'
                     }
                 }